Implementation of a Class Diagram Editor
The application consists of an area to draw the diagram and a button to add classes to the diagram. The possible actions allowed in the tool will be: inserting, moving, and deleting classes, as well as creating associations between classes.

At the bottom of the window, there are two counters indicating the number of classes and associations currently present in the represented diagram.

About Classes:

Classes will be represented by rectangles divided into three areas: the top area to show the class name, the middle area for attributes, and the bottom area for operations. Each time a class is inserted into the diagram, it will display the name as the string "Class n", where n is the creation number.
The attributes area should display the text "attributes," and the operations area will display the text "methods."
Internally, classes will be stored in a dynamic vector of class Vector, present in the java.util.* package, and located within the Diagram class. The class at position 0 of the vector will be at the back of the diagram, while the one at the last position will be on top of all others. When the user moves the mouse over a class that is semi-hidden, it should come to the front.
Classes can be moved by pressing the left mouse button on them.
To delete a class from the diagram, the user must right-click on it. If the class had any associations, those will also be deleted.
To select a class, the user must place the mouse over the class to be selected and press the "S" key. At that moment, the class's background will change from white to cyan blue. To deselect the class, the user will press the "S" key again while the cursor is over the class. Only one class can be selected in the diagram at a time. Lastly, if the "S" key is pressed when the cursor is not over any class, the last selected class will be deselected.
About Associations:

An association will be represented by a black line connecting two classes (not necessarily a straight line).
Associations can be created between two distinct classes or between a class and itself.
To create an association, the user will select one of the classes (as previously described). Once selected, the user will press the left mouse button and, without releasing it, move the cursor towards the other class (redrawing the association under the cursor). If, while dragging the association, the cursor passes over another class, that class will change color to indicate that we can create the association. If the mouse button is released while over another class, an association will be created between both classes. If the mouse button is released while the cursor is over the same class, a self-association will be created. If an association is created, no class will remain selected.
Internally, associations will be stored in a dynamic vector (java.util.Vector) within the Diagram class.