Messages broker
The functionalities of the MOM will be very similar to those of the first and second RabbitMQ tutorials. Therefore, the MOM will allow a producer program to write messages to a queue so that they can be read by a consumer program. In this version, the MOM will offer at least the following services:

Declare Queue (parameters you need): Similar to what happens in RabbitMQ, this operation will be idempotent. That is, even if it is invoked multiple times, the MOM will create a single queue with that name. Therefore, both the producer and the consumer will use this operation. You can use the Singleton pattern for this.

Publish (parameters you need): This operation is used by the producer to publish messages to a queue that it has previously declared.

Consume (parameters you need): This operation is used by the consumer to consume messages from a queue that it has previously declared.

We will keep in mind that:

If a message is published to a queue that does not exist, the message will be lost.
The consumer program will have a callback operation that the MOM will call to consume the published messages.
If there is no consumer available when a message is published, the message will remain in the queue for a maximum of 5 minutes. If no consumer has joined that queue in that time, the broker will delete the message.
If more than one consumer subscribes to the same queue, then the MOM will assign the messages using a round-robin policy, as in the second RabbitMQ tutorial.

Advanced Versions
Implement a program that lists existing queues and allows queue deletion.
Implement a mechanism for acknowledgment of processed messages, as in the second RabbitMQ tutorial.
Implement a fair dispatch policy as in the second RabbitMQ tutorial.
Implement "message durability" as in the second RabbitMQ tutorial. That is, if the MOM "crashes," the queues that have been created as durable will be recovered when the MOM is "launched" again, and the durable messages that remained in those queues will also be recovered.
The producer and consumer programs may be on different machines.